+++ lib/src/extensions/list.dart
@@ -1,3 +1,5 @@
+import 'dart:async';
+
 extension ListExtension<E> on List<E> {
   /// Split one large list to limited sub lists
   /// ```dart