+++ lib/src/extensions/map.dart
@@ -4,3 +4,17 @@ extension MapExtension<T> on Map<String, T>? {
   Map<String, T> destructure() =>
       this == null ? <String, T>{} : Map<String, T>.from(this!);
 }
+
+extension QueryParametersConversion on Map<String, dynamic> {
+  Map<String, dynamic> get asQueryParameters {
+    return map((key, value) {
+      if (value == null) {
+        return MapEntry(key, null);
+      } else if (value is Iterable) {
+        return MapEntry(key, value.map((e) => e.toString()).toList());
+      } else {
+        return MapEntry(key, value.toString());
+      }
+    });
+  }
+}